Minimize unnecessary code.
