+++ src/main/java/com/blackducksoftware/integration/phonehome/PhoneHomeClient.java
@@ -32,29 +32,35 @@ import com.blackducksoftware.integration.hub.request.Response;
 import com.blackducksoftware.integration.hub.rest.RestConnection;
 import com.blackducksoftware.integration.hub.rest.UnauthenticatedRestConnectionBuilder;
 import com.blackducksoftware.integration.log.IntLogger;
-import com.blackducksoftware.integration.phonehome.body.PhoneHomeRequestBody;
 import com.blackducksoftware.integration.phonehome.exception.PhoneHomeException;
 import com.blackducksoftware.integration.phonehome.google.analytics.GoogleAnalyticsConstants;
 import com.blackducksoftware.integration.phonehome.google.analytics.GoogleAnalyticsRequestHelper;
 import com.blackducksoftware.integration.util.CIEnvironmentVariables;
+import com.google.gson.Gson;
 
 public class PhoneHomeClient {
     public static final String SKIP_PHONE_HOME_VARIABLE = "BLACKDUCK_SKIP_PHONE_HOME";
+
     private final IntLogger logger;
     private final String googleAnalyticsTrackingId;
     private final String phoneHomeBackendUrl;
     private final int timeout;
     private final ProxyInfo proxyInfo;
     private final boolean alwaysTrustServerCertificate;
+    private final Gson gson;
 
     public PhoneHomeClient(final IntLogger logger, final String googleAnalyticsTrackingId, final int timeout, final ProxyInfo proxyInfo, final boolean alwaysTrustServerCertificate) {
+        this(logger, googleAnalyticsTrackingId, timeout, proxyInfo, alwaysTrustServerCertificate, new Gson());
+    }
+
+    public PhoneHomeClient(final IntLogger logger, final String googleAnalyticsTrackingId, final int timeout, final ProxyInfo proxyInfo, final boolean alwaysTrustServerCertificate, final Gson gson) {
         this.logger = logger;
         this.googleAnalyticsTrackingId = googleAnalyticsTrackingId;
         this.phoneHomeBackendUrl = GoogleAnalyticsConstants.BASE_URL + GoogleAnalyticsConstants.COLLECT_ENDPOINT;
         this.timeout = timeout;
         this.proxyInfo = proxyInfo;
         this.alwaysTrustServerCertificate = alwaysTrustServerCertificate;
-
+        this.gson = gson;
     }
 
     public void postPhoneHomeRequest(final PhoneHomeRequestBody phoneHomeRequestBody, final CIEnvironmentVariables environmentVariables) throws PhoneHomeException {
@@ -75,7 +81,7 @@ public class PhoneHomeClient {
         builder.setAlwaysTrustServerCertificate(alwaysTrustServerCertificate);
         final RestConnection restConnection = builder.build();
 
-        final GoogleAnalyticsRequestHelper requestHelper = new GoogleAnalyticsRequestHelper(googleAnalyticsTrackingId, phoneHomeRequestBody);
+        final GoogleAnalyticsRequestHelper requestHelper = new GoogleAnalyticsRequestHelper(gson, googleAnalyticsTrackingId, phoneHomeRequestBody);
         final Request request = requestHelper.createRequest(phoneHomeBackendUrl);
 
         try (Response response = restConnection.executeRequest(request)) {
